pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "utilities"
version = "0.1.0"
authors = [{ name = "GU Orbit Software Team", email = "uog.orbit@gmail.com" }]
requires-python = ">=3.10"
description = "A package containing utilities for GU Orbit Software"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy >= 1.24.0",
    "rasterio >= 1.3.6",
    "Pillow >= 9.4.0",
    "tensorflow >= 2.10",
    "toml >= 0.10.2",
]

[tool.setuptools]
packages = ["utilities"]

[tool.pytest.ini_options]
markers = [
    "staging: Mark a test as part of the staging environment",
    "production: Mark a test as part of the production environment",
]

[project.optional-dependencies]
dev = [
    "pytest >= 7.2.2",
    "pytest-cov >= 4.0.0",
    "pytest-mock >= 3.10.0",
    "twine >= 4.0.0",
    "Sphinx >= 6.1.3",
    "furo >= 2022.12.7",
]
